Move union to row coersion logic to hive-file-formats
